chemokine receptor-CXCR:CCR-signaling sim/requirements.txt
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.17.0
streamlit>=1.28.0
scikit-learn>=1.3.0
networkx>=3.1
pyyaml>=6.0
